pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "industrial-mcp"
version = "0.1.0"
description = "Open-source MCP server for industrial equipment - Bridge AI and your factory"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "Industrial MCP Team", email = "contact@industrial-mcp.io"}
]
keywords = [
    "mcp", "model-context-protocol", "industrial", "iot", 
    "edge-computing", "predictive-maintenance", "ai",
    "modbus", "opcua", "mqtt", "industry4.0"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Manufacturing",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: System :: Hardware",
]
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "pymodbus>=3.6.0",
    "asyncua>=1.0.0",
    "paho-mqtt>=2.0.0",
    "pyyaml>=6.0.0",
    "structlog>=24.1.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "pre-commit>=3.6.0",
]
raspberry = [
    "RPi.GPIO>=0.7.0",
    "gpiozero>=2.0.0",
]
jetson = [
    "Jetson.GPIO>=2.1.0",
    "onnxruntime-gpu>=1.16.0",
]
llm = [
    "ollama>=0.1.0",
    "chromadb>=0.4.0",
]
all = [
    "industrial-mcp[dev,raspberry,llm]",
]

[project.scripts]
industrial-mcp = "industrial_mcp.cli:main"

[project.urls]
Homepage = "https://industrial-mcp.io"
Documentation = "https://docs.industrial-mcp.io"
Repository = "https://github.com/YOUR_USERNAME/industrial-mcp"
Issues = "https://github.com/YOUR_USERNAME/industrial-mcp/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 88
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.10"
strict = true
